Add a function to create structure from a format string, similar to printf() [PERFECTIVE]
The new function packs a Bencoded structure by taking a format string and
a list of values as variable arguments. It works similarly to
ben_decode_printed(), but allows the string to values specifiers which
are replaced with values given as arguments. A specifier begins with
a percent (%) that follows a string of specifier.